# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Vallmark Gift Articles API
  version: 1.0.0
  description: "Consolidated OpenAPI (Swagger) specification generated from FastAPI\
    \ route modules under backend/routes.\nNotes:\n  - Schemas mirror pydantic models\
    \ in models.py (some descriptive text abbreviated for brevity).\n  - Generic envelope\
    \ responses use APIResponse / PaginatedResponse.\n  - Authorization: OAuth2 password\
    \ flow (token endpoint /api/auth/login) returning bearer JWT.\n"
servers:
- url: /
...
security:
- bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2Password:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/auth/login
          scopes: {}
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/auth/login
  schemas:
...
    UserRole:
      type: string
      enum:
      - customer
      - admin
      - super_admin
      - salesperson
      - store_admin
      - sales_manager
      - store_owner
      - support_executive
      - marketing_manager
    UserBase:
      type: object
      required:
      - email
      - username
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        full_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
...
    UserResponse:
      allOf:
      - $ref: '#/components/schemas/UserBase'
      - type: object
        properties:
          id:
            type: string
          email_verified:
            type: boolean
          last_login:
            type: string
            format: date-time
            nullable: true
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/UserResponse'
...
paths:
...
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: OAuth2 password login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
              - username
              - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
...